requirements/base.txt
Django>=5.1,<5.2
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
PyMuPDF>=1.24
pandas>=2.2
openpyxl>=3.1
djangorestframework>=3.15
django-filter>=24.0
drf-spectacular>=0.27
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
